+++ src/pathsim/blocks/__init__.py
@@ -1,16 +1,23 @@
 from .differentiator import *
 from .integrator import *
 from .multiplier import *
+from .converters import *
+from .comparator import *
+from .samplehold import *
 from .amplifier import *
 from .function import *
 from .spectrum import *
 from .sources import *
 from .wrapper import *
+from .counter import *
+from .filters import *
 from .switch import *
 from .adder import *
 from .scope import *
 from .delay import *
+from .noise import *
 from .ctrl import *
 from .lti import *
 from .ode import *
 from .rng import *
+from .fir import *
