--- src/pathsim/blocks/mixed/__init__.py
@@ -1,6 +1,0 @@
-from .comparator import *
-from .samplehold import *
-from .converters import *
-from .counter import *
-from .sources import *
-from .fir import *
