--- src/pathsim/blocks/rf/__init__.py
@@ -1,3 +1,0 @@
-from .sources import *
-from .filters import *
-from .noise import *
